+++ Products/CMFPlone/browser/navigation.py
@@ -154,14 +154,12 @@ class CatalogNavigationTabs(BrowserView):
             if sortOrder is not None:
                 query['sort_order'] = sortOrder
 
-        if navtree_properties.getProperty('enable_wf_state_filtering', False):
-            query['review_state'] = navtree_properties.getProperty(
-                                                    'wf_states_to_show', [])
+        if navigation_settings.filter_on_workflow:
+            query['review_state'] = navigation_settings.workflow_states_to_show
 
         query['is_default_page'] = False
 
-        if self.site_properties.getProperty('disable_nonfolderish_sections',
-                                            False):
+        if not navigation_settings.nonfolderish_tabs:
             query['is_folderish'] = True
 
         return query