+++ Products/CMFPlone/tests/testPortalCreation.py
@@ -19,6 +19,8 @@ from Products.CMFCore.permissions import AccessInactivePortalContent
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone import setuphandlers
 from Products.CMFPlone.factory import _DEFAULT_PROFILE
+from Products.CMFPlone.interfaces import INavigationSchema
+from Products.CMFPlone.interfaces import ISearchSchema
 from Products.CMFPlone.UnicodeSplitter import Splitter, I18NNormalizer
 from Products.GenericSetup.browser.manage import ExportStepsView
 from Products.GenericSetup.browser.manage import ImportStepsView
@@ -32,6 +34,7 @@ from plone.portlets.interfaces import IPortletAssignmentMapping
 from plone.portlets.interfaces import IPortletManager
 from plone.portlets.interfaces import ILocalPortletAssignmentManager
 from plone.portlets.constants import CONTEXT_CATEGORY as CONTEXT_PORTLETS
+from plone.registry.interfaces import IRegistry
 
 
 class TestPortalCreation(PloneTestCase.PloneTestCase, WarningInterceptor):
@@ -130,11 +133,6 @@ class TestPortalCreation(PloneTestCase.PloneTestCase, WarningInterceptor):
         mailhost = self.portal.MailHost
         self.assertEqual(mailhost.meta_type, 'Mail Host')
 
-    def testUseFolderContentsProperty(self):
-        # The use_folder_contents site property should be emtpy
-        props = self.portal.portal_properties.site_properties
-        self.assertEqual(props.getProperty('use_folder_contents'), ())
-
     def testFolderEditActionHasEditTitle(self):
         # Edit tab of folders should be named 'edit', not 'properties'
         folder = self.types.getTypeInfo('Folder')
@@ -179,8 +177,8 @@ class TestPortalCreation(PloneTestCase.PloneTestCase, WarningInterceptor):
 
     def testVisibleIdsProperties(self):
         # visible_ids should be a site property and a memberdata property
-        self.assertTrue(
-            self.properties.site_properties.hasProperty('visible_ids'))
+        registry = getUtility(IRegistry)
+        self.assertTrue('plone.visible_ids' in registry)
         self.assertTrue(self.memberdata.hasProperty('visible_ids'))
 
     def testFormToolTipsProperty(self):
@@ -195,8 +193,10 @@ class TestPortalCreation(PloneTestCase.PloneTestCase, WarningInterceptor):
         self.assertTrue(self.properties.navtree_properties.hasProperty('sortOrder'))
         self.assertTrue(self.properties.navtree_properties.hasProperty('sitemapDepth'))
         self.assertTrue(self.properties.navtree_properties.hasProperty('showAllParents'))
-        self.assertTrue(self.properties.navtree_properties.hasProperty('wf_states_to_show'))
-        self.assertTrue(self.properties.navtree_properties.hasProperty('enable_wf_state_filtering'))
+
+        registry = getUtility(IRegistry)
+        self.assertTrue('plone.workflow_states_to_show' in registry)
+        self.assertTrue('plone.filter_on_workflow' in registry)
 
     def testSitemapAction(self):
         # There should be a sitemap action
@@ -213,13 +213,10 @@ class TestPortalCreation(PloneTestCase.PloneTestCase, WarningInterceptor):
 
     def testUnfriendlyTypesProperty(self):
         # We should have an types_not_searched property
-        self.assertTrue(
-            self.properties.site_properties.hasProperty('types_not_searched')
-        )
-        self.assertTrue(
-            'Plone Site' in
-            self.properties.site_properties.getProperty('types_not_searched')
-        )
+        registry = getUtility(IRegistry)
+        settings = registry.forInterface(ISearchSchema, prefix="plone")
+        self.assertTrue('plone.types_not_searched' in registry)
+        self.assertTrue('Plone Site' in settings.types_not_searched)
 
     def testNonDefaultPageTypes(self):
         # We should have a default_page_types property
@@ -366,12 +363,15 @@ class TestPortalCreation(PloneTestCase.PloneTestCase, WarningInterceptor):
                          'Topic'):
             self.assertTrue(metaType in types)
 
-    def testDisableFolderSectionsSiteProperty(self):
-        # The disable_folder_sections site property should be emtpy
-        props = self.portal.portal_properties.site_properties
-        self.assertTrue(
-            props.getProperty('disable_folder_sections', None) is not None)
-        self.assertFalse(props.getProperty('disable_folder_sections'))
+    def testGenerateTabsSiteProperty(self):
+        # The generate_tabs site property should be emtpy
+        registry = getUtility(IRegistry)
+        navigation_settings = registry.forInterface(
+            INavigationSchema,
+            prefix="plone"
+        )
+        self.assertTrue('plone.generate_tabs' in registry)
+        self.assertTrue(navigation_settings.generate_tabs)
 
     def testSelectableViewsOnFolder(self):
         views = self.portal.portal_types.Folder.getAvailableViewMethods(None)
@@ -426,9 +426,9 @@ class TestPortalCreation(PloneTestCase.PloneTestCase, WarningInterceptor):
             'string:${folder_url}/view')
 
     def testEnableLivesearchProperty(self):
-        # site_properties should have enable_livesearch property
-        self.assertTrue(self.properties.site_properties \
-            .hasProperty('enable_livesearch'))
+        # registry should have enable_livesearch property
+        registry = getUtility(IRegistry)
+        self.assertTrue('plone.enable_livesearch' in registry)
 
     def testRedirectLinksProperty(self):
         self.assertTrue(self.properties.site_properties \
@@ -439,10 +439,9 @@ class TestPortalCreation(PloneTestCase.PloneTestCase, WarningInterceptor):
         self.assertEqual(self.types.Link.default_view, 'link_redirect_view')
 
     def testTTWLockableProperty(self):
-        self.assertTrue(self.properties.site_properties \
-            .hasProperty('lock_on_ttw_edit'))
-        self.assertEqual(True,
-                          self.properties.site_properties.lock_on_ttw_edit)
+        registry = getUtility(IRegistry)
+        self.assertTrue('plone.lock_on_ttw_edit' in registry)
+        self.assertEqual(True, registry['plone.lock_on_ttw_edit'])
 
     def testPortalFTIIsDynamicFTI(self):
         # Plone Site FTI should be a DynamicView FTI
@@ -876,7 +875,10 @@ class TestPortalCreation(PloneTestCase.PloneTestCase, WarningInterceptor):
                                 if r['selected']])
 
     def testNonFolderishTabsProperty(self):
-        self.assertEqual(False, self.properties.site_properties.disable_nonfolderish_sections)
+        registry = getUtility(IRegistry)
+        navigation_settings = registry.forInterface(INavigationSchema,
+                                                    prefix="plone")
+        self.assertEqual(True, navigation_settings.nonfolderish_tabs)
 
     def testNoDoubleGenericSetupImportSteps(self):
         view = ImportStepsView(self.setup, None)